More refactoring to move new command into its own class.
